Moved the getCurrentUser from api/auth to util for better clarification. Added logout functionality coresponding with login method
