+++ client/src/components/ProtectedRoute.jsx
@@ -1,6 +1,6 @@
 import { Navigate, Outlet } from "react-router-dom";
-import { getCurrentUser } from "../api/auth"; // Get user from localStorage
 import EmailVerificationMessage from "./EmailVerificationMessage";
+import getCurrentUser from "../util/getCurrentUser.js"; // Get user from localStorage
 const ProtectedRoute = () => {
     const currentUser = getCurrentUser();
     const username = currentUser ? currentUser.username : null;
