+++ client/src/util/getCurrentUser.js
@@ -1,0 +1,4 @@
+export default function getCurrentUser(){
+    const user = localStorage.getItem("user");
+    return user ? JSON.parse(user) : null;
+};
